add orchestration config data model
